move average method arg checker to utils
