+++ src/torchmetrics/functional/clustering/utils.py
@@ -20,6 +20,16 @@ from typing_extensions import Literal
 from torchmetrics.utilities.checks import _check_same_shape
 
 
+def _validate_average_method_arg(
+    average_method: Literal["min", "geometric", "arithmetic", "max"] = "arithmetic"
+) -> None:
+    if average_method not in ("min", "geometric", "arithmetic", "max"):
+        raise ValueError(
+            "Expected argument `average_method` to be one of  `min`, `geometric`, `arithmetic`, `max`,"
+            f"but got {average_method}"
+        )
+
+
 def calculate_entropy(x: Tensor) -> Tensor:
     """Calculate entropy for a tensor of labels.
